// src/app/core/enums/nr-id.enum.ts
/**
 * Enum de NRs, para IDs únicos e consistentes.
 */
export enum NRIdEnum {
  NR01 = "NR01",
  NR04 = "NR04",
  NR05 = "NR05",
  NR06 = "NR06",
  NR10 = "NR10",
  NR11 = "NR11",
  NR12 = "NR12",
  NR13 = "NR13",
  NR15 = "NR15",
  NR17 = "NR17",
  NR18 = "NR18",
  NR20 = "NR20",
  NR23 = "NR23",
  NR33 = "NR33",
  NR35 = "NR35",
  ALL = "ALL",
}